chore: refactor images folder
