+++ src/features/pages/home/our-platforms/data.tsx
@@ -5,7 +5,7 @@ import DTraderIcon from 'images/svg/dtrader/dtrader_icon2.svg'
 import DMT5Icon from 'images/svg/dmt5/dmt5_icon2.svg'
 import DerivGOIcon from 'images/svg/deriv-go/deriv_go_icon.svg'
 import DerivXIcon from 'images/svg/deriv-x/deriv_x_icon.svg'
-import DBotIcon from 'images/svg/dbot/deriv_bot_icon.svg'
+import DBotIcon from 'images/svg/dbot/dbot_icon5.svg'
 import BinaryBotIcon from 'images/svg/binarybot_icon.svg'
 import SmartTraderIcon from 'images/svg/smart_trader_icon.svg'
 import DerivEzIcon from 'images/svg/deriv-ez/deriv_ez_icon.svg'
