+++ src/images/svg/dbot/index.ts
@@ -5,6 +5,6 @@ export { default as DbotBg } from './dbot_bg.svg'
 export { default as BotBannerLogo } from './dbot_banner_logo.svg'
 export { default as DtIcon } from '../dtrader/dtrader_icon.svg'
 export { default as DxIcon } from '../deriv-x/deriv_x_icon.svg'
-export { default as EzIcon } from './dbot_ez_icon.svg'
+export { default as EzIcon } from '../deriv-ez/deriv_ez_icon3.svg'
 export { default as GoIcon } from '../deriv-go/deriv_go_icon.svg'
-export { default as Mt5Icon } from './mt5.svg'
+export { default as Mt5Icon } from '../dmt5/dmt5_icon4.svg'
