+++ src/images/svg/deriv-ez/index.ts
@@ -1,8 +1,8 @@
-export { default as DbIcon } from './deriv_ez_db_icon.svg'
+export { default as DbIcon } from '../dbot/dbot_icon4.svg'
 export { default as DerivEzLogoRed } from './deriv_ez_logo_red.svg'
 export { default as DerivEzLogo } from './deriv_ez_logo.svg'
 export { default as DerivEzQR } from './deriv_ez_qr.svg'
-export { default as Dmt5Icon } from './deriv_ez_dmt5_icon.svg'
+export { default as Dmt5Icon } from '../dmt5/dmt5_icon5.svg'
 export { default as DtIcon } from '../dtrader/dtrader_icon.svg'
 export { default as DxIcon } from '../deriv-x/deriv_x_icon.svg'
 export { default as EasyToStart } from './deriv_ez_easy_to_start.svg'
