+++ src/images/svg/deriv-x/index.ts
@@ -1,11 +1,11 @@
 import Icon24_7 from 'images/svg/deriv-x/deriv_x_24_7.svg'
 import IconMultipleMarketing from 'images/svg/deriv-x/deriv_x_multiple_marketing.svg'
 import IconNewPromising from 'images/svg/deriv-x/deriv_x_new_promising.svg'
-import MT5Icon from 'images/svg/deriv-x/deriv_x_mt5.svg'
-import DTIcon from 'images/svg/deriv-x/deriv_x_dt_icon.svg'
-import EZIcon from 'images/svg/deriv-x/deriv_x_ez_icon.svg'
+import MT5Icon from 'images/svg/dmt5/dmt5_icon4.svg'
+import DTIcon from 'images/svg/dtrader/dtrader_icon4.svg'
+import EZIcon from 'images/svg/deriv-ez/deriv_ez_icon3.svg'
 import DGoIcon from 'images/svg/deriv-go/deriv_go_icon.svg'
-import DBIcon from 'images/svg/deriv-x/deriv_x_db_icon.svg'
+import DBIcon from 'images/svg/dbot/dbot_icon3.svg'
 import DerivXLogo from 'images/svg/deriv-x/deriv_x.svg'
 import DerivXBannerLogo from 'images/svg/deriv-x/deriv_x_banner_logo.svg'
