+++ src/pages/deriv-go/_other-apps.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import mt5Icon from '../../images/svg/dmt5/dmt5_logo2.svg'
-import dtIcon from '../../images/svg/deriv-go/deriv_go_dt_icon.svg'
-import ezIcon from '../../images/svg/deriv-go/deriv_go_ez_icon.svg'
+import dtIcon from '../../images/svg/dtrader/dtrader_icon4.svg'
+import ezIcon from '../../images/svg/deriv-ez/deriv_ez_icon3.svg'
 import dxIcon from '../../images/svg/deriv-x/deriv_x_icon.svg'
-import dbIcon from '../../images/svg/deriv-go/deriv_go_db_icon.svg'
+import dbIcon from '../../images/svg/dbot/dbot_icon3.svg'
 import GenericCarousel, { TRenderableData } from 'components/custom/carousel/_platforms-carousel'
 
 const carouselData: TRenderableData[] = [
