+++ src/pages/dmt5/_other-platforms.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import dxIcon from '../../images/svg/deriv-x/deriv_x_icon.svg'
 import dtIcon from '../../images/svg/dtrader/dtrader_icon.svg'
-import ezIcon from '../../images/svg/dmt5/dmt5_ez_icon.svg'
+import ezIcon from '../../images/svg/deriv-ez/deriv_ez_icon3.svg'
 import dGoIcon from '../../images/svg/deriv-go/deriv_go_icon.svg'
-import dbIcon from '../../images/svg/dmt5/dmt5_db_icon.svg'
+import dbIcon from '../../images/svg/dbot/dbot_icon4.svg'
 import GenericCarousel, { TRenderableData } from 'components/custom/carousel/_platforms-carousel'
 
 const carouselData: TRenderableData[] = [
