+++ src/pages/trading-specification/components/_available-platform.tsx
@@ -4,7 +4,7 @@ import useRegion from 'components/hooks/use-region'
 import { Flex } from 'components/containers'
 import DerivMT5 from 'images/common/markets/mt5.svg'
 import DerivX from 'images/svg/trading-specification/trading_spec_deriv_x.svg'
-import DerivEZ from 'images/svg/dtrader/dtrader_ez_icon.svg'
+import DerivEZ from 'images/svg/deriv-ez/deriv_ez_icon4.svg'
 import { Localize, LocalizedLink, localize } from 'components/localization'
 import device from 'themes/device'
 import { Header } from 'components/elements'
